Refaktoryzacja części odpowiedzialnej za rekordy w pamięci. Refaktoryzacja niepotrzebnych rzutowań. Zmiana kilku nazw w projekcie.

+++ LabZKT/Main/Simulation/BitTextBox.cs
@@ -34,7 +34,7 @@ namespace LabZKT
             loc.X = x;
             loc.Y = y;
             Location = loc;
-            Font = new System.Drawing.Font("Tahoma", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(238)));
+            Font = new System.Drawing.Font("Tahoma", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, 238);
         }
 
         protected override void OnKeyPress(KeyPressEventArgs e)
