+++ LabZKT/Main/Simulation/NumericTextBox.cs
@@ -28,7 +28,7 @@ namespace LabZKT
             loc.Y = y;
             Location = loc;
             BackColor = Color.White;
-            Font = new System.Drawing.Font("Tahoma", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(238)));
+            Font = new System.Drawing.Font("Tahoma", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, 238);
         }
 
         protected override void OnKeyPress(KeyPressEventArgs e)
